add >NAME tag, use wire for symbol outline
Using wire for the symbol outline makes it look
more like other schematic symbols (appears red
rather than gray.) Having >NAME that can be changed
is good if there are multiple parts in the schematic.
I think I also moved the symbol so that the origin
is in the center, which makes it easier to drag around,
select, etc.